requirements.txt
discord.py>=2.3.2
requests==2.33.1
beautifulsoup4==4.14.3
python-dotenv==1.0.0
cryptography>=41.0.0
